+++ tests/process_message_test.py
@@ -1,7 +1,6 @@
-# To write tests here
 import unittest
 import mock
-from plugins.food_bot_plugin import process_message, Responses, outputs
+from plugins.food_bot_plugin import process_message
 
 
 class TestProcessMessage(unittest.TestCase):
@@ -94,14 +93,3 @@ class TestProcessMessage(unittest.TestCase):
         self.assertFalse(mock_response.rate.called)
         self.assertFalse(mock_response.enter_comment.called)
         self.assertFalse(mock_response.get_average_ratings.called)
-
-
-class TestHelpFunction(unittest.TestCase):
-
-    def test_help_returns_correct_string(self):
-        Responses.show_help(u'D0CLHTW20')
-
-        self.assertEqual(len(outputs), 1)
-        self.assertEqual(type(outputs[0]), list)
-        self.assertEqual(outputs[0][0], u'D0CLHTW20')
-        self.assertEqual(outputs[0][1], Responses.help_text)
